Binary Search For finding Occurances of Number
The binary search approach is used to find the total no. of occurances or frequencies of a given no. In this approach first the array should be in ascending order (sorted) then we check the first time that number is started using a FirstOcc() function, in this a mid is find and if the no. is smaller than the left side of mid is searched only and a recursive approach is used to decrease the value of high pointer. And the reverse goes for the finding the last occurances in the given array. Then by subtracting both the last occ by first occ we find out desired output by increasing 1 to the result.