Sort 0s,1s,2s
In this problem we are given an array of 0s, 1s and 2s and we have to sort it in ascending order.
So a good approach would be to find the total number of each 0s, 1s and 2s and use a individual for loop for each 0s, 1s and 2s. The range of first for loop will be equal to the total number of 0s present in that array and in the new array put all the 0. Same goes for all the loops

Another best approach could be by using Dutch National Flag Algorithm , In this for each values we use a case in switch. If 0 is found then low is swapped with mid and low and mid is increased by 1. If 1 is found only mid is increased. If 2 is found then mid is swapped with high and only high is increased.