Update frequencyOfNumber.java
some changes made i.e removed for loop that checks every elements is not equal to the given number. Instead a simple things is checked i.e. if the given number is greater than or equal to the length of frequency array then the result would be zero else the result would be given by that index value of frequency.