Create UnionAndIntersectionOfTwoArrays.java
In this we are provided with two arrays and we use TreeSet collections to store non repeated values in Union method, then the values are stored in arrayList and returned. For intersection method we insert the values of only one array and iterate over other array and check if that treeset contains that interated value of another array and then store it in arrayList and return it.